% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_input.R
\name{get_pkg_data}
\alias{get_pkg_data}
\title{Get package data}
\usage{
get_pkg_data(package)
}
\arguments{
\item{package}{name of package (a character vector)}
}
\value{
list of all datasets in package
}
\description{
Get package data
}
\examples{
require(tidyr)
get_pkg_data("tidyr")
}
