% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_ui}
\alias{mod_data_input_ui}
\title{Import data UI module}
\usage{
mod_data_input_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny UI module
}
\description{
Import data UI module
}
