% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plot.R
\name{mod_plot_server}
\alias{mod_plot_server}
\title{Plot server module}
\usage{
mod_plot_server(id, plot_inputs)
}
\arguments{
\item{id}{module id}

\item{scatter_inputs}{inputs from mod_var_select}
}
\value{
shiny server module
}
\description{
Plot server module
}
