% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_points_output.R, R/mod_scatter_output.R
\name{mod_scatter_output_server}
\alias{mod_scatter_output_server}
\title{Scatter plot server module}
\usage{
mod_scatter_output_server(id, var_inputs, app_data)

mod_scatter_output_server(id, var_inputs, app_data)
}
\arguments{
\item{id}{module id}

\item{var_inputs}{inputs from mod_var_input}

\item{app_data}{data for application}
}
\value{
shiny server module

shiny server module
}
\description{
Scatter plot server module

Plot variables Server module
}
