% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_points_output.R, R/mod_scatter_output.R
\name{mod_scatter_output_ui}
\alias{mod_scatter_output_ui}
\title{Scatter plot UI module}
\usage{
mod_scatter_output_ui(id)

mod_scatter_output_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny UI module

shiny UI module
}
\description{
Scatter plot UI module

Plot variables UI module
}
