% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_select.R
\name{mod_var_select_server}
\alias{mod_var_select_server}
\title{Select variables Server module}
\usage{
mod_var_select_server(id, input_data)
}
\arguments{
\item{id}{module id}

\item{input_data}{pkg data from \code{mod_data_input}}
}
\value{
shiny server module
}
\description{
Select variables Server module
}
\section{Use}{

The returned list includes:
\itemize{
\item \code{df} is the same object that is stored in \code{pkg_data()}, which is built
from the \code{input$pkg} and \code{input$ds}.
\item \code{x}: the x axis variable
\item \code{y}: the y axis variable
\item \code{col}: the color variable
\item \code{facet}: the facet variable
\item \code{alpha}: the alpha (opacity) for the points
\item \code{size}: the size of the points
}
}

