% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsap-package.R
\docType{package}
\name{tsap-package}
\alias{tsap}
\alias{tsap-package}
\title{tsap: Test Your Shiny Package App server functions}
\description{
This shiny app-package was developed to demonstrate testing with testthat and shiny's testServer().
}
\keyword{internal}
