% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPkgApp-package.R
\docType{package}
\name{testServerAppPkg-package}
\alias{testServerAppPkg}
\alias{testServerAppPkg-package}
\title{testServerAppPkg: Test Your Shiny Package App server functions}
\description{
This shiny app was developed as a package using devtools and usethis.
}
\keyword{internal}
