% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_input.R
\name{check_pkg}
\alias{check_pkg}
\title{Check if package is loaded, if not load it}
\usage{
check_pkg(pkg)
}
\arguments{
\item{pkg}{name of package (a character vector)}
}
\value{
Package: 'name' loaded or Loading package: 'name'
}
\description{
If package is not installed, install with \code{install.packages(dependencies = TRUE)}
}
\examples{
check_pkg("plotly")
}
