% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_server}
\alias{mod_data_input_server}
\title{Import data server module}
\usage{
mod_data_input_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny server module
}
\description{
Import data server module
}
\section{Use}{


\code{input$data} includes a call to \code{shiny::observe()},
and \code{shiny::bindEvent(input$pkg)} to ensure both values are
included in the returned object.
}

