% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_input.R
\name{mod_data_input_server}
\alias{mod_data_input_server}
\title{Import data server module}
\usage{
mod_data_input_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
shiny server module
}
\description{
Import data server module
}
