% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_scatter.R
\name{mod_scatter_server}
\alias{mod_scatter_server}
\title{Scatter plot server module}
\usage{
mod_scatter_server(id, vars, df)
}
\arguments{
\item{id}{module id}

\item{vars}{inputs from mod_var_input}

\item{df}{data for application}
}
\value{
shiny server module
}
\description{
Scatter plot server module
}
